requirements.txt
# Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# Deep Learning
torch==2.0.1
torchvision==0.15.2
timm==0.9.12

# Image Processing
Pillow==10.0.1
opencv-python==4.8.1.78

# Visualization
matplotlib==3.8.0
numpy==1.24.3

# Additional Dependencies
scipy==1.11.3